requirements.txt
asynctest
click
coverage
nose
pylint
python-daemon
rednose
requests

